Address #22, #27, #26

+++ fabric-version1x/removeNetwork.sh
@@ -26,19 +26,33 @@
 #####################################################################################################################
 # Code                                                                                                              #
 #####################################################################################################################
+dockerimages=$(docker ps -aq)
+if [[ -n "$dockerimages" ]]; 
+then
+    ### Stop all docker containers
+    docker stop ns_prometheus
+    docker rm ns_prometheus
+    cd peerOperation/
+    docker-compose -f docker_compose_peer_cli_couchdb_deoni.yaml down
+    cd ../orderingService/
+    docker-compose -f docker_compose_orderer_unibw.yaml down
+    cd ../applications/
+    docker-compose -f docker_compose_cli_unibw.yaml down
+    
+    docker stop $dockerimages
+    ### Remove all docker containers
+    docker rm $dockerimages
+    ### Remove old chaincode images
+    docker rmi $(docker images | grep dev | tr -s ' ' | cut -d ' ' -f 3)
+    cd ..
+else
+    echo "No docker images to delete..." 
+fi
 
-### Stop all docker containers
-docker stop $(docker ps -aq)
 
-### Remove all docker containers
-docker rm $(docker ps -aq)
-
-### Remove old chaincode images
-docker rmi $(docker images | grep dev | tr -s ' ' | cut -d ' ' -f 3)
-
-### Remove all files from configTransactions
+### Remove all files from configTransactions but the README file
 cd configTransactions/
 shopt -s extglob
-sudo rm !(README.md)
+rm -f !(README.md)
 
 echo "See you next time! 👋"
